Refactoring: Separate files related with auto_scaler
