+++ lib/ecs_deploy.rb
@@ -1,7 +1,6 @@
 require "ecs_deploy/version"
 require "ecs_deploy/configuration"
 
-require 'aws-sdk-ec2'
 require 'aws-sdk-ecs'
 require 'logger'
 require 'terminal-table'
