+++ lib/ecs_deploy/auto_scaler.rb
@@ -1,8 +1,9 @@
-require 'aws-sdk-autoscaling'
-require 'aws-sdk-cloudwatch'
-require 'yaml'
-require 'logger'
-require 'time'
+require "logger"
+require "time"
+require "yaml"
+
+require "ecs_deploy/auto_scaler/auto_scaling_config"
+require "ecs_deploy/auto_scaler/service_config"
 
 module EcsDeploy
   module AutoScaler
@@ -125,310 +126,6 @@ module EcsDeploy
         end
 
         @logger.debug "Stop #{name}"
-      end
-    end
-
-    module ConfigBase
-      def initialize(attributes = {}, logger)
-        attributes.each do |key, val|
-          send("#{key}=", val)
-        end
-        @logger = logger
-      end
-
-      def logger
-        @logger
-      end
-    end
-
-    SERVICE_CONFIG_ATTRIBUTES = %i(name cluster region auto_scaling_group_name step max_task_count min_task_count idle_time scheduled_min_task_count cooldown_time_for_reach_max upscale_triggers downscale_triggers desired_count)
-    ServiceConfig = Struct.new(*SERVICE_CONFIG_ATTRIBUTES) do
-      include ConfigBase
-
-      MAX_DETACHABLE_INSTANCE_COUNT = 20
-
-      def initialize(attributes = {}, logger)
-        super
-        self.idle_time ||= 60
-        self.max_task_count = Array(max_task_count)
-        self.upscale_triggers = upscale_triggers.to_a.map do |t|
-          TriggerConfig.new(t.merge(region: region), logger)
-        end
-        self.downscale_triggers = downscale_triggers.to_a.map do |t|
-          TriggerConfig.new(t.merge(region: region), logger)
-        end
-        self.max_task_count.sort!
-        self.desired_count = fetch_service.desired_count
-        @reach_max_at = nil
-        @last_updated_at = nil
-      end
-
-      def client
-        Aws::ECS::Client.new(
-          access_key_id: EcsDeploy.config.access_key_id,
-          secret_access_key: EcsDeploy.config.secret_access_key,
-          region: region,
-          logger: logger
-        )
-      end
-
-      def idle?
-        return false unless @last_updated_at
-
-        diff = Process.clock_gettime(Process::CLOCK_MONOTONIC, :second) - @last_updated_at
-        diff < idle_time
-      end
-
-      def current_min_task_count
-        return min_task_count if scheduled_min_task_count.nil? || scheduled_min_task_count.empty?
-
-        scheduled_min_task_count.find(-> { {"count" => min_task_count} }) { |s|
-          from = Time.parse(s["from"])
-          to = Time.parse(s["to"])
-          (from..to).cover?(Time.now)
-        }["count"]
-      end
-
-      def overheat?
-        return false unless @reach_max_at
-        (Process.clock_gettime(Process::CLOCK_MONOTONIC, :second) - @reach_max_at) > cooldown_time_for_reach_max
-      end
-
-      def fetch_service
-        res = client.describe_services(cluster: cluster, services: [name])
-        raise "Service \"#{name}\" is not found" if res.services.empty?
-        res.services[0]
-      rescue => e
-        AutoScaler.error_logger.error(e)
-      end
-
-      def update_service(difference)
-        next_desired_count = desired_count + difference
-        current_level = max_task_level(desired_count)
-        next_level = max_task_level(next_desired_count)
-        if current_level < next_level && overheat? # next max
-          level = next_level
-          @reach_max_at = nil
-          AutoScaler.logger.info "Service \"#{name}\" is overheat, uses next max count"
-        elsif current_level < next_level && !overheat? # wait cooldown
-          level = current_level
-          now = Process.clock_gettime(Process::CLOCK_MONOTONIC, :second)
-          @reach_max_at ||= now
-          AutoScaler.logger.info "Service \"#{name}\" waits cooldown elapsed #{(now - @reach_max_at).to_i}sec"
-        elsif current_level == next_level && next_desired_count >= max_task_count[current_level] # reach current max
-          level = current_level
-          now = Process.clock_gettime(Process::CLOCK_MONOTONIC, :second)
-          @reach_max_at ||= now
-          AutoScaler.logger.info "Service \"#{name}\" waits cooldown elapsed #{(now - @reach_max_at).to_i}sec"
-        elsif current_level == next_level && next_desired_count < max_task_count[current_level]
-          level = current_level
-          @reach_max_at = nil
-          AutoScaler.logger.info "Service \"#{name}\" clears cooldown state"
-        elsif current_level > next_level
-          level = next_level
-          @reach_max_at = nil
-          AutoScaler.logger.info "Service \"#{name}\" clears cooldown state"
-        end
-
-        cl = client
-        next_desired_count = [next_desired_count, max_task_count[level]].min
-        cl.update_service(
-          cluster: cluster,
-          service: name,
-          desired_count: next_desired_count,
-        )
-        cl.wait_until(:services_stable, cluster: cluster, services: [name]) do |w|
-          w.before_wait do
-            AutoScaler.logger.debug "wait service stable [#{name}]"
-          end
-        end if difference < 0
-        @last_updated_at = Process.clock_gettime(Process::CLOCK_MONOTONIC, :second)
-        self.desired_count = next_desired_count
-        AutoScaler.logger.info "Update service \"#{name}\": desired_count -> #{next_desired_count}"
-      rescue => e
-        AutoScaler.error_logger.error(e)
-      end
-
-      def fetch_container_instances_in_cluster
-        arns = []
-        cl = client
-        resp = cl.list_container_instances(cluster: cluster)
-        resp.each do |r|
-          arns.concat(r.container_instance_arns)
-        end
-
-        chunk_size = 50
-        container_instances = []
-        arns.each_slice(chunk_size) do |arn_chunk|
-          is = cl.describe_container_instances(cluster: cluster, container_instances: arn_chunk).container_instances
-          container_instances.concat(is)
-        end
-
-        container_instances
-      end
-
-      def fetch_container_instance_arns_in_service
-        arns = []
-        resp = client.list_container_instances(cluster: cluster, filter: "task:group == service:#{name}")
-        resp.each do |r|
-          arns.concat(r.container_instance_arns)
-        end
-
-        arns
-      end
-
-      private
-
-      def max_task_level(count)
-        max_task_count.index { |i| count <= i } || max_task_count.size - 1
-      end
-    end
-
-    TriggerConfig = Struct.new(:alarm_name, :region, :state, :step) do
-      include ConfigBase
-
-      def client
-        Aws::CloudWatch::Client.new(
-          access_key_id: EcsDeploy.config.access_key_id,
-          secret_access_key: EcsDeploy.config.secret_access_key,
-          region: region,
-          logger: logger
-        )
-      end
-
-      def match?
-        fetch_alarm.state_value == state
-      end
-
-      def fetch_alarm
-        res = client.describe_alarms(alarm_names: [alarm_name])
-
-        raise "Alarm \"#{alarm_name}\" is not found" if res.metric_alarms.empty?
-        res.metric_alarms[0].tap do |alarm|
-          AutoScaler.logger.debug("#{alarm.alarm_name} state is #{alarm.state_value}")
-        end
-      rescue => e
-        AutoScaler.error_logger.error(e)
-      end
-    end
-
-    AutoScalingConfig = Struct.new(:name, :region, :buffer) do
-      include ConfigBase
-
-      def client
-        Aws::AutoScaling::Client.new(
-          access_key_id: EcsDeploy.config.access_key_id,
-          secret_access_key: EcsDeploy.config.secret_access_key,
-          region: region,
-          logger: logger
-        )
-      end
-
-      def ec2_client
-        Aws::EC2::Client.new(
-          access_key_id: EcsDeploy.config.access_key_id,
-          secret_access_key: EcsDeploy.config.secret_access_key,
-          region: region,
-          logger: logger
-        )
-      end
-
-      def instances(reload: false)
-        if reload || @instances.nil?
-          resp = client.describe_auto_scaling_groups({
-            auto_scaling_group_names: [name],
-          })
-          @instances = resp.auto_scaling_groups[0].instances
-        else
-          @instances
-        end
-      end
-
-      def update_auto_scaling_group(total_service_count, service_config)
-        desired_capacity = total_service_count + buffer.to_i
-
-        current_asg = client.describe_auto_scaling_groups({
-          auto_scaling_group_names: [name],
-        }).auto_scaling_groups[0]
-
-        if current_asg.desired_capacity > desired_capacity
-          diff = current_asg.desired_capacity - desired_capacity
-          container_instance_arns_in_service = service_config.fetch_container_instance_arns_in_service
-          container_instances_in_cluster = service_config.fetch_container_instances_in_cluster
-          deregisterable_instances = container_instances_in_cluster.select do |i|
-            i.pending_tasks_count == 0 && !running_essential_task?(i, container_instance_arns_in_service)
-          end
-
-          AutoScaler.logger.info "Fetch deregisterable instances: #{deregisterable_instances.map(&:ec2_instance_id).inspect}"
-
-          deregistered_instance_ids = []
-          deregisterable_instances.each do |i|
-            break if deregistered_instance_ids.size >= diff
-            begin
-              service_config.client.deregister_container_instance(cluster: service_config.cluster, container_instance: i.container_instance_arn, force: true)
-              deregistered_instance_ids << i.ec2_instance_id
-            rescue Aws::ECS::Errors::InvalidParameterException
-            end
-          end
-
-          AutoScaler.logger.info "Deregistered instances: #{deregistered_instance_ids.inspect}"
-
-          detach_and_terminate_instances(deregistered_instance_ids)
-
-          AutoScaler.logger.info "Update auto scaling group \"#{name}\": desired_capacity -> #{desired_capacity}"
-        elsif current_asg.desired_capacity < desired_capacity
-          client.update_auto_scaling_group(
-            auto_scaling_group_name: name,
-            min_size: 0,
-            max_size: [current_asg.max_size, desired_capacity].max,
-            desired_capacity: desired_capacity,
-          )
-          AutoScaler.logger.info "Update auto scaling group \"#{name}\": desired_capacity -> #{desired_capacity}"
-        end
-      rescue => e
-        AutoScaler.error_logger.error(e)
-      end
-
-      def detach_and_terminate_instances(instance_ids)
-        return if instance_ids.empty?
-
-        instance_ids.each_slice(MAX_DETACHABLE_INSTANCE_COUNT) do |ids|
-          client.detach_instances(
-            auto_scaling_group_name: name,
-            instance_ids: ids,
-            should_decrement_desired_capacity: true
-          )
-        end
-
-        AutoScaler.logger.info "Detach instances from ASG #{name}: #{instance_ids.inspect}"
-        sleep 3
-
-        ec2_client.terminate_instances(instance_ids: instance_ids)
-
-        AutoScaler.logger.info "Terminated instances: #{instance_ids.inspect}"
-      rescue => e
-        AutoScaler.error_logger.error(e)
-      end
-
-      def detach_and_terminate_orphan_instances(service_config)
-        container_instance_ids = service_config.fetch_container_instances_in_cluster.map(&:ec2_instance_id)
-        orphans = instances(reload: true).reject { |i| container_instance_ids.include?(i.instance_id) }.map(&:instance_id)
-
-        return if orphans.empty?
-
-        targets = ec2_client.describe_instances(instance_ids: orphans).reservations[0].instances.select do |i|
-          (Time.now - i.launch_time) > 600
-        end
-
-        detach_and_terminate_instances(targets.map(&:instance_id))
-      rescue => e
-        AutoScaler.error_logger.error(e)
-      end
-
-      def running_essential_task?(instance, container_instance_arns_in_service)
-        return false if instance.running_tasks_count == 0
-
-        container_instance_arns_in_service.include?(instance.container_instance_arn)
       end
     end
   end
