+++ lib/ecs_deploy/auto_scaler/config_base.rb
@@ -1,0 +1,16 @@
+module EcsDeploy
+  module AutoScaler
+    module ConfigBase
+      def initialize(attributes = {}, logger)
+        attributes.each do |key, val|
+          send("#{key}=", val)
+        end
+        @logger = logger
+      end
+
+      def logger
+        @logger
+      end
+    end
+  end
+end
